Converts to eslint and starts the cleanup and util expansion
Adds an isDefined and isUndefined function to the utils module for
easy checking.

Cleans up according to the eslint rules in several places and adds
an eslint (lint) task to the gulp file. Right now it is not part of
the build because of the number of errors. Someday it will be added
once that is more under control.